+++ quicksort.h
@@ -2,6 +2,6 @@
 #define QUICKSORT_H
 #include <stddef.h>
 
-void quicksort(unsigned *A, int n);
+void quicksort(unsigned *A, size_t n);
 
 #endif // #ifndef QUICKSORT_H
